CM-1597: Extracted OnlineExperimentBuilderImpl into separate file to de-clutter OnlineExperimentImpl.
